+++ src/internal/operations/inject/inject.go
@@ -28,6 +28,12 @@ type (
 
 		IsServiceEnableder
 
+		// CollectMetadata returns a list of metadata collections. In
+		// case of service that have just a single underlying service like
+		// OneDrive or SharePoint, it will mostly just have a single
+		// collection per manifest reason, but in the case of groups, it
+		// will contain a collection each for the underlying service,
+		// for example one per SharePoint site.
 		CollectMetadata(
 			ctx context.Context,
 			r inject.RestoreProducer,
